chore: debug tests
